+++ trunk/source/menu/menu_config.cpp
@@ -32,7 +32,7 @@ void CMenu::_hideConfig(bool instant)
 	m_btnMgr.hide(m_configLblCfg4, instant);
 	m_btnMgr.hide(m_configBtnCfg4, instant);
 	for(u8 i = 0; i < ARRAY_SIZE(m_configLblUser); ++i)
-		if(m_configLblUser[i] != (u16)-1)
+		if(m_configLblUser[i] != -1)
 			m_btnMgr.hide(m_configLblUser[i], instant);
 }
 
@@ -67,7 +67,7 @@ void CMenu::_showConfig(void)
 			partitionname[i] = toupper(partitionname[i]);
 
 		for(u8 i = 0; i < ARRAY_SIZE(m_configLblUser); ++i)
-			if(m_configLblUser[i] != (u16)-1)
+			if(m_configLblUser[i] != -1)
 				m_btnMgr.show(m_configLblUser[i]);
 		
 		m_btnMgr.setText(m_configLblPartition, (string)partitionname);
