+++ trunk/source/menu/menu_home.cpp
@@ -5,21 +5,21 @@
 #include "loader/nk.h"
 #include "loader/sys.h"
 
-u32 m_homeLblTitle;
-u32 m_exittoLblTitle;
+s16 m_homeLblTitle;
+s16 m_exittoLblTitle;
 
-u32 m_homeBtnSettings;
-u32 m_homeBtnReloadCache;
-u32 m_homeBtnUpdate;
-u32 m_homeBtnHelp;
-u32 m_homeBtnAbout;
-u32 m_homeBtnExitTo;
+s16 m_homeBtnSettings;
+s16 m_homeBtnReloadCache;
+s16 m_homeBtnUpdate;
+s16 m_homeBtnHelp;
+s16 m_homeBtnAbout;
+s16 m_homeBtnExitTo;
 
-u32 m_homeBtnExitToHBC;
-u32 m_homeBtnExitToMenu;
-u32 m_homeBtnExitToPriiloader;
-u32 m_homeBtnExitToBootmii;
-u32 m_homeBtnExitToNeek;
+s16 m_homeBtnExitToHBC;
+s16 m_homeBtnExitToMenu;
+s16 m_homeBtnExitToPriiloader;
+s16 m_homeBtnExitToBootmii;
+s16 m_homeBtnExitToNeek;
 
 STexture m_homeBg;
